Fix the complexity issue of removing dead decorate
Change the container type of decorate instructions from std:;vector to a
combination of std::list and std::unordered_set/map. This avoid the
copy overhead during eliminateDeadDecorations(), and keep the
instruction dumping order unchanged.

Issue: may be an overkill for now, and the new type should be a
template.