ParseHelper: do not use default precision for build-in functions
This removes default precision qualifiers from parsing the build-ins. This is needed for functions,
so they get precision computed from their arguments (that is already handled in glslang).

Without that all FS build-in functions other than pack/unpack and sampler/image operations are performed in mediump.

Additionally, all build-in variables are checked to have explicit precision qualifier in declaration.