HLSL: step 2: throw big switch to use HLSL parser for intrinsics.
This commit switches over to creating HLSL intrinsics in HLSL syntax,
where previously they had been created in GLSL syntax.

This also changes the return qualifier on the intrinsics from global to
temp, which changes a bunch of test results.  That is the only test output
change.  This is in its own commit, to keep it separate.