HLSL: disallow promotion of texture/image/etc types for atomics and sample methods.
If there is (say) a signed int buffer type, then the signed version of the atomic
operation should be used.  Previously, this was generating an ambiguous overload
error, because it was trying to convert the buffer resource itself.  Now, if
the buffer (tx/etc) is a uint, but int parameters are given, then the params
are converted rather than trying to use the wrong overload.

New test in hlsl.promote.atomic.frag