Remove linker output for every stage of every program
The presence of non-warning/error output is bad since the callers have
to filter out warning/error messages with regular expressions or the
like to keep messages that are actually important.

TIntermediate::warn/error prefix their messages with the stage the issue
is with so this message appears completely redundant.