Refactor TType::contains* methods (nonfunctional)
There were many (~8) different places in TType which all knew how to traverse
the struct/type hierarchy.  There's a need to add another, but I didn't
want to duplicate the traversal code again.  This is a small refactoring
which passes a predicate to a single traverse-and-test method.  That also
shortens all the containsSomething() methods from 9 lines of body to 1.

There are no test differences: it's nonfunctional.