Using C++11 features for replacing current OS specific code.
TLS code remove at all, replaced by the thread_local specifier. This also
impacts in the initialization process simplifying it a lot. The
functions Init/DetachProcess, Init/DetachThread have been removed. The
only function needed to be called for a proper initializations is
InitializeMemoryPools (FreeMemoryPools is optional during th shutdown).

TLS Support for MSVC 2013 is limited, only PODs are allowed to be used
and the specific extension should be used for it. This prevents us to do
further code simplifications.

The OGLCompilersDLL dynamic library is not needed anymore and has been
removed from the solution.

The way how is handle the PoolAllocators has been slightly modified. The
SetThreadPoolAllocator now receives a unique_ptr for the new
PoolAllocator and returns another unique_ptr for the old one. The goal
is to clearly define who is the ownership of those pool allocators in any
moment. Therefore the member variables pools in any object must be from
now a unique_ptr.

Overall this change removes 560 lines and adds 87.