Fix memory leak of the default thread pool allocator
FreeGlobalPools did not get called in order to free the pool allocator
allocated in InitializeMemoryPools. Make it so.

Callers to SetThreadPoolAllocator are already responsible for freeing
their newly-set allocator. Since we now call FreeGlobalPools, this may
result in a double-free as callers are not keen on restoring the
previously set allocator after use.

Workaround this by keeping a reference to the default allocator and
delete it in FreeGlobalPools.