Fix memory leak of pragma table in TIntermAggregate
AST nodes are allocated from the thread pool and deleted in bulk. This
is achieved by overriding the new and delete operators in TIntermNode
(via the POOL_ALLOCATOR_NEW_DELETE macro).

The overriding delete operator does not call the class' destructor,
which TIntermAggregate implements in order to delete its TPragmaTable
instance, resulting in a memory leak.

Fix this by having TPragmaTable instances be allocated from the thread
pool.