SPV: Do not generate ImageMSArray capability for sampled images
According to the specification description for the capability:

"ImageMSArray: An MS operand in OpTypeImage indicates multisampled,
 used without a sampler."

So the capability must not be included when the only images
used by the shader are sampled. The description of the Sampled
member of OpTypeImage is:

"Sampled indicates whether or not this image will be accessed
 in combination with a sampler, and must be one of the
 following values:
 0 indicates this is only known at run time, not at compile time
 1 indicates will be used with sampler
 2 indicates will be used without a sampler (a storage image)"

So we do not want to generate this capability when the image
Sampled value is 1.

There are a few tests that seemed to expect the capability to be
generated when only sampled images are used. The patch updates these
tests accordingly.