WIP: HLSL: add methods to track user structure in texture return type.
Incomplete and not yet functional.  This will turn into support for user structure returns
from textures.

Some languages allow a restricted set of user structure types returned from texture sampling
operations.  Restrictions include the total vector size of all components may not exceed 4,
and the basic types of all members must be identical.

This adds underpinnings for that ability.  Because storing a whole TType or even a simple
TTypeList in the TSampler would be expensive, the structure definition is packed into a
small bitfield, which can represent (1) a scalar or vector type as before, or (2) structures
abiding the necessary restrictions.

The details of the storage are protected.  There are abstraction methods accepting and
returning a TType (such as may have been parsed from a grammar).  The abstraction methods
turn the type into the internal representation, or vice versa.

Currently this is hooked up enough to do what it used to do for vectors and scalars.
Places in the code which used to directly set or get the vector size now use the new
abstraction.  The new abstraction cannot yet handle the structure, but that's a matter of
fleshing it out, and will be added soon.