SPV: Fix GLSL/HLSL atomicDecrement.
Make atomic decrement return the post decrement, and
bump the generator number.

This is done separately from other changes coming from GitHub to
reduce the differences that will be seen when doing that.
(This commit changes every single SPV output result.)

This represents both the following commits in GitHub master:
07ed11f9a08820505eef1e128bd758b8b9b5ed5b
48d6e798bc76b8e58be844a8cef8bc6f843a909e