SPV/GL: allow shift operation for int<<uint of different rank
From GLSL spec:
"The shift operators (<<) and (>>). For both operators, the operands
must be signed or unsigned integers or integer vectors. One operand
can be signed while the other is unsigned. In all cases, the resulting
type will be the same type as the left operand. If the first operand
is a scalar, the second operand has to be a scalar as well. If the
first operand is a vector, the second operand must be a scalar or a
vector, and the result is computed component-wise. The result is
undefined if the right operand is negative, or greater than or equal
to the number of bits in the left expression's base type"

It clearly points that int/uint combinations are allowed.

From SPIR-V spec:
"The type of each Base and Shift must be a scalar or vector of integer
type. Base and Shift must have the same number of components. The
number of components and bit width of the type of Base must be the
same as in Result Type.  Shift is treated as unsigned. The result is
undefined if Shift is greater than the bit width of the components of
Base."

It clearly points that the requirement for rank is for the Base and
Result, not on the Base and Shift.

v2: simplifies check. Add test.

Fixes:
https://github.com/KhronosGroup/glslang/issues/1296