Added a [[dynamically_uniform]] variable attribute.
Developers are constantly asking for a way to make `Uniform` decorations
appear in the output SPIR-V. I've taken a stab at reusing the attribute
mechanism introduced in the control flow GLSL extension to add a variable
attribute [[dynamically_uniform]] that can apply to variables, function
parameters, and function return types - which results in the corresponding
variables having the `Uniform` decoration.

I haven't drafted any GLSL extension for a couple of reasons:

1. Do we actually need one for functionality that is unconditionally in
SPIR-V anyway?
2. Even if I do require an extension, it'll be trivial to draft it after
the technical approach I've taken is approved.

I've also added extensive warnings for cases when a user might be moving
a value of unknown uniformity into a value that is denoted as dynamically
uniform - a normal cause of errors.