[GLSL][PP] extension pp check after non-pp tokens
**[Purpose]**

Currently we lack of checking whether "#extension" is used after any non-pp tokens/statements.

According to spec, this situation should be treated as an error.

Here is a fix for issues with following pattern:

**[Shader]**
...

void main{
...
}

GLSL should notify users that this would lead to a compilation error as "#extension" should not be used after non-pp tokens.

**[Concerns]**
I don't know whether it is OK to use a simple flag/mask attribute to represent current pp state.
I will add tests if this change is OK according to your concerns.