Add implementation dependent limit for uniform buffer bindings
GLSL could correctly check uniform bindings for atomic counter and samplers. But failing to check binding limit for uniform block.

As other two cases use built-in MAX variables for checking and to provide to users to fetch and check, a new limit variable for uniform block could be used for limit checking and also shader coding usage.

GLSL doesn't have a built-in level variable for this, so we don't set it as a built-in, but an implementation dependent limit.

PS: As this bingding limit depends on hardware implementation, we only
set INT32_MAX as test data in StandAlone tools.