Use one element to represent the last member of SSBO which is unsized array
Purpose:
The size of precedingMemberUnsizedArray which is unsized array wasn't calculated.
Sulotion: Add unsized array size into block size.

Shader Type: Comp.

Reference:
According to Spec:
For the property BUFFER_DATA_SIZE, the implementation-dependent minimum
total buffer object size is written to params. This value is the size, in basic
machine units, required to hold all active variables associated with an active uniform
block, shader storage block, or atomic counter buffer. If the final member of
an active shader storage block is an array with no declared size, the minimum buffer
size is computed assuming the array was declared as an array with one element.