Update Constant.cpp
Purpose:
Fix calculation of log2 and exp2.

Modification:
During test, the result of 'const float n = log2(14.0 + float(int(-3.0) - int(3.0)));' is incorrect, the expression is fold to constant at compile-time, but glslang fold the log2() to 1.44 * log(), which made the result incorrect, in fact the system has provided such built-in functions, Also exp2 has the same issues, so the issue about exp2 is also fixed in this change.