Force SPIRV project to be a static lib
This project is dependent on glslang which is dependent on this project. One of these needs to be static to avoid external linker errors when BUILD_SHARED_LIBS = ON
This reverses the dependency order for the SPIRV and glslang libraries.

This allows glslang to compile as a shared library and fixes compiling BUILD_SHARED_LIBS = ON