Change const char* to std::string for extBuiltinMap in GlslangToSpv
extBuiltinMap is defined as the map <libName, libId>. const char* as
the key is not correct. Actually, we care the content of the string
rather than pointer value. The problem is not exposed because we always
use pre-defined const string spv::E_SPV_XXX. If the usage is changed
to this, the problem of redefinitions occurs because getExtBuiltins()
is malfunction.

  std::string ext1("XXX");
  getExtBuiltIns(ext1.c_str());

  std:;string ext2("XXX");
  getExtBuiltins(ext2.c_str());