glslang: Only export public interface for SOs
Default to `-fvisibility=hidden`, and annotate the public glslang
interface with `GLSLANG_EXPORT` to change the visibility of these
cherry-picked symbols to default. This is also used by Windows
builds for `__declspec(dllexport)`-ing the public DLL interface.

This allows us to classify API changes into those that are
publicly backwards compatible, and those that are not.

Note that libSPIRV will likely need similar treatment.