Fix precision propagation through function calls
The TIntermTyped::propagatePrecision was given a precision and it
propagated it to its operands.  For aggregates, it was applying the
precision to all its operands.  This is not correct for function calls
(including built-ins), and this change limits this to only constructors.

For example, glslang was mistranslating the following code:

    mediump sampler2D s;
    highp ivec2 coords = ...;
    mediump result = texture(s, vec3(int(coords.x) & 3, int(coords.y) & y, 3));

where parameter 1 of the texture call is a constructor (which doesn't
have a precision per spec), and the expressions inside the constructor
don't have a precision either (because they are either constructors or
literals).  glslang was incorrectly applying the precision of the result
(mediump) to all the parameters, including the coordinate calculation,
causing visual artifacts in ANGLE traces.

Fixes: #2740